+++ frontend/src/store/index.js
@@ -1,6 +1,7 @@
 import { configureStore } from '@reduxjs/toolkit'
 import authReducer from './slices/authSlice'
 import walletReducer from './slices/walletSlice'
+import diamondReducer from './slices/diamondSlice'
 import gameReducer from './slices/gameSlice'
 import rankReducer from './slices/rankSlice'
 
@@ -8,6 +9,7 @@ const store = configureStore({
   reducer: {
     auth: authReducer,
     wallet: walletReducer,
+    diamond: diamondReducer,
     game: gameReducer,
     rank: rankReducer,
   },